Profile page connected to backend partially
