+++ cmput404-group-project/src/components/Navbar.tsx
@@ -10,6 +10,7 @@ function Navbar() {
   const closeMobileMenu = () => setClick(false);
   let signInSignOut;
   let homePageFeed;
+  let profile;
   React.useEffect(() => {
     if (localStorage.getItem('user') != null) {
       setUser(JSON.parse(localStorage.getItem('user')!));
@@ -88,6 +89,22 @@ function Navbar() {
               </NavLink>
 
             </li>
+            {user !== null ?
+              <li className="nav-item">
+
+                <NavLink
+                  to="/profile"
+                  className={({ isActive }) =>
+                    "nav-links" + (isActive ? " activated" : "")
+                  }
+                  onClick={closeMobileMenu}
+                  style={{ textDecoration: 'none' }}
+                >
+                  Profile
+                </NavLink>
+
+              </li>
+              : null}
             <li className="nav-item">
               {signInSignOut}
             </li>
